requirements.txt
python-dotenv      
Pillow             
openai             
keyboard           
sv-ttk             